feat: 实现加入房间功能

+++ src/Model/PlayerInfo.cs
@@ -10,5 +10,5 @@ public class PlayerInfo : UserInfo
     [JsonIgnore]
     public Client? client;
 
-    public short dealerWind;
+    public byte dealerWind;
 }
